connector_test: Move the mock to the same test script
Move the mockSTS to the script test that uses it, to keep the
homogeneity with the others test scripts regarding the mocks that the
define and use.
