+++ connector_test.go
@@ -8,9 +8,23 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials"
+	"github.com/aws/aws-sdk-go/aws/request"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/sts"
+	"github.com/aws/aws-sdk-go/service/sts/stsiface"
 )
+
+type mockSTS struct {
+	stsiface.STSAPI
+
+	// Mocking of GetCallerIdentity
+	gcio   *sts.GetCallerIdentityOutput
+	gcierr error
+}
+
+func (m mockSTS) GetCallerIdentityWithContext(_ aws.Context, _ *sts.GetCallerIdentityInput, _ ...request.Option) (*sts.GetCallerIdentityOutput, error) {
+	return m.gcio, m.gcierr
+}
 
 func TestConnector_setRegion(t *testing.T) {
 	var ec2Regions = &ec2.DescribeRegionsOutput{
